[Python API] Bind test_utils
