+++ inference-engine/ie_bridges/python/CMakeLists.txt
@@ -66,10 +66,6 @@ set (PYTHON_BRIDGE_SRC_ROOT ${CMAKE_CURRENT_SOURCE_DIR})
 add_subdirectory (src/openvino/inference_engine)
 add_subdirectory (src/openvino/offline_transformations)
 
-if (ENABLE_TESTS)
-    add_subdirectory(src/openvino/test_utils)
-endif()
-
 if(ENABLE_WHEEL)
     add_subdirectory(wheel)
 endif()
