--- inference-engine/ie_bridges/python/src/openvino/test_utils/__init__.py
@@ -1,7 +1,0 @@
-# -*- coding: utf-8 -*-
-# Copyright (C) 2018-2021 Intel Corporation
-# SPDX-License-Identifier: Apache-2.0
-
-from .test_utils_api import *
-
-__all__ = ['CompareNetworks']
