--- inference-engine/ie_bridges/python/src/openvino/test_utils/test_utils_api_impl_defs.pxd
@@ -1,11 +1,0 @@
-# Copyright (C) 2018-2021 Intel Corporation
-# SPDX-License-Identifier: Apache-2.0
-
-from libcpp cimport bool
-from libcpp.string cimport string
-from libcpp.pair cimport pair
-
-from ..inference_engine.ie_api_impl_defs cimport IENetwork
-
-cdef extern from "test_utils_api_impl.hpp" namespace "InferenceEnginePython":
-    cdef pair[bool, string] CompareNetworks(IENetwork lhs, IENetwork rhs)
