+++ src/bindings/python/src/openvino/runtime/test_utils/__init__.py
@@ -1,0 +1,5 @@
+# -*- coding: utf-8 -*-
+# Copyright (C) 2021 Intel Corporation
+# SPDX-License-Identifier: Apache-2.0
+
+from openvino.pyopenvino.test_utils import compare_functions
