+++ src/inference/dev_api/src/pyopenvino/test_utils/CMakeLists.txt
@@ -1,4 +1,4 @@
-# Copyright (C) 2018-2021 Intel Corporation
+# Copyright (C) 2021 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
 #
 
@@ -9,24 +9,19 @@ set(CMAKE_ARCHIVE_OUTPUT_DIRECTORY ${PYTHON_BRIDGE_OUTPUT_DIRECTORY}/test_utils)
 set(CMAKE_COMPILE_PDB_OUTPUT_DIRECTORY ${PYTHON_BRIDGE_OUTPUT_DIRECTORY}/test_utils)
 set(CMAKE_PDB_OUTPUT_DIRECTORY ${PYTHON_BRIDGE_OUTPUT_DIRECTORY}/test_utils)
 
-file(GLOB SOURCES ${CMAKE_CURRENT_SOURCE_DIR}/test_utils_api.pyx
-                  ${CMAKE_CURRENT_SOURCE_DIR}/test_utils_api_impl.cpp
-                  ${CMAKE_CURRENT_SOURCE_DIR}/test_utils_api_impl.hpp
-                  ${CMAKE_CURRENT_SOURCE_DIR}/test_utils_api_impl_defs.pxd)
-
-set_source_files_properties(${CMAKE_CURRENT_SOURCE_DIR}/test_utils_api.pyx
-                            PROPERTIES CYTHON_IS_CXX ON)
+file(GLOB SOURCES ${CMAKE_CURRENT_SOURCE_DIR}/test_utils.cpp
+        ${CMAKE_CURRENT_SOURCE_DIR}/test_utils.hpp)
 
 # create target
 
-cython_add_module(${TARGET_NAME} ${SOURCES})
+pybind11_add_module(${TARGET_NAME} MODULE ${SOURCES})
 
-add_dependencies(${TARGET_NAME} ie_api)
+add_dependencies(${TARGET_NAME} pyopenvino)
 ov_python_disable_intel_warnings(${TARGET_NAME})
 
 if(COMMAND ie_add_vs_version_file)
     ie_add_vs_version_file(NAME ${TARGET_NAME}
-                           FILEDESCRIPTION "Test Utils Python library")
+            FILEDESCRIPTION "Test Utils Python library")
 endif()
 
 if(InferenceEngineDeveloperPackage_FOUND)
@@ -36,7 +31,6 @@ else()
 endif()
 
 target_include_directories(${TARGET_NAME} PRIVATE "${CMAKE_CURRENT_SOURCE_DIR}")
-target_include_directories(${TARGET_NAME} SYSTEM PRIVATE "${CMAKE_CURRENT_SOURCE_DIR}/../inference_engine")
 target_link_libraries(${TARGET_NAME} PRIVATE ${link_libraries})
 
 # Compatibility with python 2.7 which has deprecated "register" specifier
@@ -46,12 +40,12 @@ endif()
 
 # perform copy
 add_custom_command(TARGET ${TARGET_NAME}
-    POST_BUILD
-    COMMAND ${CMAKE_COMMAND} -E copy ${PYTHON_BRIDGE_SRC_ROOT}/src/openvino/test_utils/__init__.py ${CMAKE_LIBRARY_OUTPUT_DIRECTORY}/__init__.py
-)
+        POST_BUILD
+        COMMAND ${CMAKE_COMMAND} -E copy  ${CMAKE_CURRENT_LIST_DIR}/../openvino/test_utils/__init__.py ${CMAKE_LIBRARY_OUTPUT_DIRECTORY}/__init__.py
+        )
 
 add_clang_format_target(${TARGET_NAME}_clang FOR_TARGETS ${TARGET_NAME}
-                        EXCLUDE_PATTERNS ".*\\.cxx;.*\\.pxd;.*\\.pyx")
+        EXCLUDE_PATTERNS ".*\\.cxx")
 
 # install
 
@@ -60,8 +54,3 @@ install(TARGETS ${TARGET_NAME}
         COMPONENT tests EXCLUDE_FROM_ALL
         LIBRARY DESTINATION ${PYTHON_BRIDGE_CPACK_PATH}/${PYTHON_VERSION}/openvino/test_utils
         COMPONENT tests EXCLUDE_FROM_ALL)
-
-install(PROGRAMS __init__.py
-        DESTINATION ${PYTHON_BRIDGE_CPACK_PATH}/${PYTHON_VERSION}/openvino/test_utils
-        COMPONENT tests
-        EXCLUDE_FROM_ALL)
