+++ src/inference/dev_api/src/pyopenvino/test_utils/test_utils.hpp
@@ -1,0 +1,11 @@
+// Copyright (C) 2021 Intel Corporation
+// SPDX-License-Identifier: Apache-2.0
+//
+
+#pragma once
+
+#include <pybind11/pybind11.h>
+
+namespace py = pybind11;
+
+void regmodule_test_utils(py::module m);
